Move util.py to varspec.py
