+++ metasyn/var.py
@@ -10,7 +10,7 @@ import polars as pl
 from metasyn.distribution.base import BaseDistribution
 from metasyn.privacy import BasePrivacy, BasicPrivacy
 from metasyn.provider import BaseDistributionProvider, DistributionProviderList
-from metasyn.util import DistributionSpec
+from metasyn.varspec import DistributionSpec
 
 
 class MetaVar():
